// src/gen/Flight_pb.d.ts
export declare type FlightDescriptor = Message<"arrow.flight.protocol.FlightDescriptor"> & {
  /**
   * @generated from field: arrow.flight.protocol.FlightDescriptor.DescriptorType type = 1;
   */
  type: FlightDescriptor_DescriptorType;

  /**
   *
   * Opaque value used to express a command. Should only be defined when
   * type = CMD.
   *
   * @generated from field: bytes cmd = 2;
   */
  cmd: Uint8Array;

  /**
   *
   * List of strings identifying a particular dataset. Should only be defined
   * when type = PATH.
   *
   * @generated from field: repeated string path = 3;
   */
  path: string[];
};

/**
 * Describes the message arrow.flight.protocol.FlightDescriptor.
 * Use `create(FlightDescriptorSchema)` to create a new message.
 */
export declare const FlightDescriptorSchema: GenMessage<FlightDescriptor>;

/**
 *
 * Describes what type of descriptor is defined.
 *
 * @generated from enum arrow.flight.protocol.FlightDescriptor.DescriptorType
 */
export enum FlightDescriptor_DescriptorType {
  /**
   * Protobuf pattern, not used.
   *
   * @generated from enum value: UNKNOWN = 0;
   */
  UNKNOWN = 0,

  /**
   *
   * A named path that identifies a dataset. A path is composed of a string
   * or list of strings describing a particular dataset. This is conceptually
   *  similar to a path inside a filesystem.
   *
   * @generated from enum value: PATH = 1;
   */
  PATH = 1,

  /**
   *
   * An opaque command to generate a dataset.
   *
   * @generated from enum value: CMD = 2;
   */
  CMD = 2,
}

/**
 * Describes the enum arrow.flight.protocol.FlightDescriptor.DescriptorType.
 */
export declare const FlightDescriptor_DescriptorTypeSchema: GenEnum<FlightDescriptor_DescriptorType>;

/**
 *
 * The access coordinates for retrieval of a dataset. With a FlightInfo, a
 * consumer is able to determine how to retrieve a dataset.
 *
 * @generated from message arrow.flight.protocol.FlightInfo
 */
export declare type FlightInfo = Message<"arrow.flight.protocol.FlightInfo"> & {
  /**
   * The schema of the dataset in its IPC form:
   *   4 bytes - an optional IPC_CONTINUATION_TOKEN prefix
   *   4 bytes - the byte length of the payload
   *   a flatbuffer Message whose header is the Schema
   *
   * @generated from field: bytes schema = 1;
   */
  schema: Uint8Array;

  /**
   *
   * The descriptor associated with this info.
   *
   * @generated from field: arrow.flight.protocol.FlightDescriptor flight_descriptor = 2;
   */
  flightDescriptor?: FlightDescriptor | undefined;

  /**
   *
   * A list of endpoints associated with the flight. To consume the
   * whole flight, all endpoints (and hence all Tickets) must be
   * consumed. Endpoints can be consumed in any order.
   *
   * In other words, an application can use multiple endpoints to
   * represent partitioned data.
   *
   * If the returned data has an ordering, an application can use
   * "FlightInfo.ordered = true" or should return all data in a
   * single endpoint. Otherwise, there is no ordering defined on
   * endpoints or the data within.
   *
   * A client can read ordered data by reading data from returned
   * endpoints, in order, from front to back.
   *
   * Note that a client may ignore "FlightInfo.ordered = true". If an
   * ordering is important for an application, an application must
   * choose one of them:
   *
   * * An application requires that all clients must read data in
   *   returned endpoints order.
   * * An application must return all data in a single endpoint.
   *
   * @generated from field: repeated arrow.flight.protocol.FlightEndpoint endpoint = 3;
   */
  endpoint: FlightEndpoint[];

  /**
   * Set these to -1 if unknown.
   *
   * @generated from field: int64 total_records = 4;
   */
  totalRecords: bigint;

  /**
   * @generated from field: int64 total_bytes = 5;
   */
  totalBytes: bigint;

  /**
   *
   * FlightEndpoints are in the same order as the data.
   *
   * @generated from field: bool ordered = 6;
   */
  ordered: boolean;

  /**
   *
   * Application-defined metadata.
   *
   * There is no inherent or required relationship between this
   * and the app_metadata fields in the FlightEndpoints or resulting
   * FlightData messages. Since this metadata is application-defined,
   * a given application could define there to be a relationship,
   * but there is none required by the spec.
   *
   * @generated from field: bytes app_metadata = 7;
   */
  appMetadata: Uint8Array;
};
// ...
export declare type Ticket = Message<"arrow.flight.protocol.Ticket"> & {
  /**
   * @generated from field: bytes ticket = 1;
   */
  ticket: Uint8Array;
};

/**
 * Describes the message arrow.flight.protocol.Ticket.
 * Use `create(TicketSchema)` to create a new message.
 */
export declare const TicketSchema: GenMessage<Ticket>;
// ...
export declare type Location = Message<"arrow.flight.protocol.Location"> & {
  /**
   * @generated from field: string uri = 1;
   */
  uri: string;
};

/**
 * Describes the message arrow.flight.protocol.Location.
 * Use `create(LocationSchema)` to create a new message.
 */
export declare const LocationSchema: GenMessage<Location>;

/**
 *
 * A particular stream or split associated with a flight.
 *
 * @generated from message arrow.flight.protocol.FlightEndpoint
 */
export declare type FlightEndpoint = Message<"arrow.flight.protocol.FlightEndpoint"> & {
  /**
   *
   * Token used to retrieve this stream.
   *
   * @generated from field: arrow.flight.protocol.Ticket ticket = 1;
   */
  ticket?: Ticket | undefined;

  /**
   *
   * A list of URIs where this ticket can be redeemed via DoGet().
   *
   * If the list is empty, the expectation is that the ticket can only
   * be redeemed on the current service where the ticket was
   * generated.
   *
   * If the list is not empty, the expectation is that the ticket can be
   * redeemed at any of the locations, and that the data returned will be
   * equivalent. In this case, the ticket may only be redeemed at one of the
   * given locations, and not (necessarily) on the current service. If one
   * of the given locations is "arrow-flight-reuse-connection://?", the
   * client may redeem the ticket on the service where the ticket was
   * generated (i.e., the same as above), in addition to the other
   * locations. (This URI was chosen to maximize compatibility, as 'scheme:'
   * or 'scheme://' are not accepted by Java's java.net.URI.)
   *
   * In other words, an application can use multiple locations to
   * represent redundant and/or load balanced services.
   *
   * @generated from field: repeated arrow.flight.protocol.Location location = 2;
   */
  location: Location[];

  /**
   *
   * Expiration time of this stream. If present, clients may assume
   * they can retry DoGet requests. Otherwise, it is
   * application-defined whether DoGet requests may be retried.
   *
   * @generated from field: google.protobuf.Timestamp expiration_time = 3;
   */
  expirationTime?: Timestamp | undefined;

  /**
   *
   * Application-defined metadata.
   *
   * There is no inherent or required relationship between this
   * and the app_metadata fields in the FlightInfo or resulting
   * FlightData messages. Since this metadata is application-defined,
   * a given application could define there to be a relationship,
   * but there is none required by the spec.
   *
   * @generated from field: bytes app_metadata = 4;
   */
  appMetadata: Uint8Array;
};