// src/gen/FlightSql_pb.d.ts
export declare type CommandGetSqlInfo = Message<"arrow.flight.protocol.sql.CommandGetSqlInfo"> & {
  /**
   *
   * Values are modelled after ODBC's SQLGetInfo() function. This information is intended to provide
   * Flight SQL clients with basic, SQL syntax and SQL functions related information.
   * More information types can be added in future releases.
   * E.g. more SQL syntax support types, scalar functions support, type conversion support etc.
   *
   * Note that the set of metadata may expand.
   *
   * Initially, Flight SQL will support the following information types:
   * - Server Information - Range [0-500)
   * - Syntax Information - Range [500-1000)
   * Range [0-10,000) is reserved for defaults (see SqlInfo enum for default options).
   * Custom options should start at 10,000.
   *
   * If omitted, then all metadata will be retrieved.
   * Flight SQL Servers may choose to include additional metadata above and beyond the specified set, however they must
   * at least return the specified set. IDs ranging from 0 to 10,000 (exclusive) are reserved for future use.
   * If additional metadata is included, the metadata IDs should start from 10,000.
   *
   * @generated from field: repeated uint32 info = 1;
   */
  info: number[];
};

/**
 * Describes the message arrow.flight.protocol.sql.CommandGetSqlInfo.
 * Use `create(CommandGetSqlInfoSchema)` to create a new message.
 */
export declare const CommandGetSqlInfoSchema: GenMessage<CommandGetSqlInfo>;
// ...
export declare type CommandGetTables = Message<"arrow.flight.protocol.sql.CommandGetTables"> & {
  /**
   *
   * Specifies the Catalog to search for the tables.
   * An empty string retrieves those without a catalog.
   * If omitted the catalog name should not be used to narrow the search.
   *
   * @generated from field: optional string catalog = 1;
   */
  catalog?: string | undefined;

  /**
   *
   * Specifies a filter pattern for schemas to search for.
   * When no db_schema_filter_pattern is provided, all schemas matching other filters are searched.
   * In the pattern string, two special characters can be used to denote matching rules:
   *    - "%" means to match any substring with 0 or more characters.
   *    - "_" means to match any one character.
   *
   * @generated from field: optional string db_schema_filter_pattern = 2;
   */
  dbSchemaFilterPattern?: string | undefined;

  /**
   *
   * Specifies a filter pattern for tables to search for.
   * When no table_name_filter_pattern is provided, all tables matching other filters are searched.
   * In the pattern string, two special characters can be used to denote matching rules:
   *    - "%" means to match any substring with 0 or more characters.
   *    - "_" means to match any one character.
   *
   * @generated from field: optional string table_name_filter_pattern = 3;
   */
  tableNameFilterPattern?: string | undefined;

  /**
   *
   * Specifies a filter of table types which must match.
   * The table types depend on vendor/implementation. It is usually used to separate tables from views or system tables.
   * TABLE, VIEW, and SYSTEM TABLE are commonly supported.
   *
   * @generated from field: repeated string table_types = 4;
   */
  tableTypes: string[];

  /**
   * Specifies if the Arrow schema should be returned for found tables.
   *
   * @generated from field: bool include_schema = 5;
   */
  includeSchema: boolean;
};

/**
 * Describes the message arrow.flight.protocol.sql.CommandGetTables.
 * Use `create(CommandGetTablesSchema)` to create a new message.
 */
export declare const CommandGetTablesSchema: GenMessage<CommandGetTables>;
// ...
export declare type CommandStatementQuery = Message<"arrow.flight.protocol.sql.CommandStatementQuery"> & {
  /**
   * The SQL syntax.
   *
   * @generated from field: string query = 1;
   */
  query: string;

  /**
   * Include the query as part of this transaction (if unset, the query is auto-committed).
   *
   * @generated from field: optional bytes transaction_id = 2;
   */
  transactionId?: Uint8Array | undefined;
};

/**
 * Describes the message arrow.flight.protocol.sql.CommandStatementQuery.
 * Use `create(CommandStatementQuerySchema)` to create a new message.
 */
export declare const CommandStatementQuerySchema: GenMessage<CommandStatementQuery>;